Livestatus: implemented option parsing and "debug"
It's now possible to pass options via nagios.cfg
to the module. In the previous version the complete
option string was taken as socket path. Now you
can pass options with the form name=value. Currently
only debug=1 is defined, which turns on logging
of queries. The socket path is specified without
a '=' (for compatibility reasons).