Livestatus: fix problems with timeperiod
Nagios has some race conditions in the function which
determines, if a time period is currently active. This
commit introduces a global timeperiod cache which is
acutalized once a minute and simply looked up as needed.
This avoid the race conditions and saves CPU time.
(would be a good idea for Nagios it self to use such
 a cache). This bug fix was sponsored by Heinz.