SNMP: Fixed missing OID detection in SNMP check processing.
There was a problem when the first column had OID gaps in the middle. This affected e.g. the cisco_locif check.
The OIDs were fetched from the first column and only appended by the other columns but the OID list for checking
was not sorted and therefor the holes could not be filled correctly.