Using the real enclosure number as check item now.
The problem with the 'enclosure device id' which was used before is
being generated somehow. So on identical hardware with identical
raid setup the 'enclosure device ids' can be different. This is a
bad behaviour when hardcoding checks for a group of hosts in check_mk.

Instead of the 'enclosure device id' the 'enclosure number' is used now.
Which is also better to identify the enclosure hardware without having
a look at the configuration.