FIX: snmp: fix detection of missing OIDs
END-OIDs are now not any longer assumed to consist of one
component but we take the complete part after the scan prefix
as END OIDs. This avoids the creation of completely empty lines
like ['', ''] when in the MIB the indices of entries consist
of more than one component (like in apc_symmetra_temp).

We are not yet sure if this patch has a negative influence
in other not-yet tested situations...