Handle None-parameter for checks with factory_settings
If a check created by inventory has None as a parameter
but for that check type exists factory settings, then
we automatically change the None into an empty dictionary
so that the factory settings will apply. This helps in
cases where an earlier version of the check works without
parameters and a later implementation uses a dictionary.