WATO: Added wato_create_users_from_htpasswd option
Users & Contacts: Added new option wato_create_users_from_htpasswd
to make handling of users only appearing in htpasswd configurable. The
default is set to True to keep the default behaviour. Users only appearing
in htpasswd will not be shown in this GUI when the option is set to
False.