Automatically remove duplicate checks on SNMP+TCP
This was very annoying. Now if a host as a TCP based and SNMP based variant
of the same check then the SNMP based one is silently dropped. It is still
handled during inventory and visible in WATO. But when the actual monitoring
configuration is being created, it is omitted.

In the rare case where the user wants to prefer the SNMP based check, he
can still use disabled_checks and disable the TCP based one for the host
in question.