veeam_client: new check to monitor status of veeam clients with special agent plugin
This new check monitors client operations of the veeam backup tool. It comes with the
check script, template for pnp4nagios, and a special windows agent plugin which needs
to be installed on the veeam server. It monitors the state of the backup operation,
and stores the total backup size, backup duration and data thruput of the backup as
performance data.