FIX Improved error handling of SNMP commands
In the past, in normal SNMP mode, the stderr output of all
snmp related commands was completely suppressed. It was hard
to find out the source of problems in some cases, for example
timeouts. The execution of those programs has been changed to
output stderr messages for walks in all cases and for gets only
when -v is given.