FIX More verbose error output for SNMP errors on the command line
In the past, in non-inline SNMP mode, the stderr output of all
SNMP related commands was completely suppressed. It was hard
to find out the source of problems in some cases, for example
timeouts. The execution of those programs has been changed to
output stderr messages for walks in all cases and for gets only
when -v is given.