FIX Fix exception in WATO host editor on custom tag without topic
When you've defined a host tag group in WATO that has no topic,
then the host edit dialog would be interrupted with an exception.
A simple workaround to this is to alwasy define a topic.

The problem has now been fixed. It had shortly been introduced by fixing a
problem appearing when you used a slash in the name of a tag group.