FIX Old log entries were shown in event history first
In previous versions, when the event history views had a limit
applied, like it is configured by default, it happened that
too old entries were read first. When reaching the limit of entries,
the processing stopped and newer events were ignored.

The order of the log files during processing has now been changed.