FIX Fix CMC not executing any Check_MK checks after config reload
CMC uses Check_MK in the <tt>--keepalive</tt> mode. That mode mixed up
it's filedescriptors after a configuration reload. That resultet in the
output of the checks not arriving at the CMC. The CMC waited and considered
the helper as busy. All Check_MK based checks got stale. A restart fixed
that.

That problem is now solved.