Changed user profiles (e.g. pw changes) are now replicated in distributed setups
When a user changed its password or its user profile attributes in previous versions, these
changes were only applied on the local (central) site in a distributed WATO setup. These
changes were only replicated to the slave sites when an administrator started replicating
other things to the slave sites. This has been changed now. When a user changes the profile
in a distributed setup, a sync dialog is opened and the sync of this user profile is
started for each slave site.