FIX Allow overriding of active checks by more specific rule
If you define an active Check via WATO in the main folder and then
want to override this check in a subfolder, that was not possible
up to now. Now if you define an active check <b>with the same
type and service description</b> by a more specific rule, then
that rule will override the more general ones (i.e. those
coming later).