if64: check now supports clustering
If you have a cluster of two or more hosts, and the check is configured
as a clustered service for an interface, the interface with the highest
output traffic is considered the active interface, and only this interface
is monitored and assigned to the cluster. The check denotes from which node
the interface data is taken.