Distributed WATO now disabled WATO on slave sites per default
When pushing a configuration to a slave site on that site WATO is
now disabled per default. This can be configured in the site
configuration with a new check box <i>Disable configuration via WATO on this site</i>,
which is check per default.

Disabling WATO on slave sites is a good idea, because any configuration
changes via WATO will be lost on the next synchronization.