Rule for checking agents for wanted version
You can now define a rule (in WATO it is called "Check for correct version of
Check_MK agent") to check if the Check_MK Agents have one wanted version.
You may give a version number manually or take the version number of the
monitoring site as a reference.

If the Agent has a different version, the Check_MK service on this host returns
a WARN status, or what ever status you did define in the rule "Status of the
Check_MK service".