FIX logwatch: fix case where rule wouldn't be applied
If you are using CMC core <tt>cmk -v</tt> on the commandline and
had inventorized the checks by a newer version of Check_MK then the
reclassification patterns of logwatch would not work. This has been
fixed. The reason was that the inventory now correctly sets <tt>None</tt>
as a parameter whereas older versions used an empty string.