FIX LDAP: Improved slightly missleading logging of LDAP sync actions
In some situations a call to WATO issued by the user are causing LDAP syncs,
for example when a user starts the sync interactively or the cache lifetime
has exceeded. In the past, the name of the triggering user was logged
to the audit / pending log, which was a bit missleading. Now, the name
of the user is not logged anymore.